% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hello_world}
\alias{hello_world}
\title{Internal smoke-test entry point used during package build.}
\usage{
hello_world()
}
\description{
Internal smoke-test entry point used during package build.
}
\keyword{internal}
