% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_find_motifs}
\alias{rs_find_motifs}
\title{Internal FFI: find motifs from R-supplied vectors. Use the
user-facing \code{\link[=find_motifs_rs]{find_motifs_rs()}} wrapper instead.}
\usage{
rs_find_motifs(
  cell_ids,
  cell_types,
  edge_source,
  edge_target,
  size,
  colored,
  anchored_on
)
}
\description{
Internal FFI: find motifs from R-supplied vectors. Use the
user-facing \code{\link[=find_motifs_rs]{find_motifs_rs()}} wrapper instead.
}
\keyword{internal}
