% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_motif_instances}
\alias{rs_motif_instances}
\title{Internal FFI: vertices of instances belonging to named motif classes.}
\usage{
rs_motif_instances(
  n_nodes,
  type_codes,
  type_levels,
  edge_source,
  edge_target,
  size,
  motif_ids,
  max_per_class
)
}
\description{
Internal FFI: vertices of instances belonging to named motif classes.
}
\keyword{internal}
